requirements.txt
accelerate==0.34.2
agentops==0.3.11
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.29.2
antlr4-python3-runtime==4.9.3
anyio==4.5.0
asttokens==2.4.1
attrs==24.2.0
av==13.0.0
azure-core==1.31.0
azure-storage-blob==12.23.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
botocore==1.35.23
build==1.2.2
CacheControl==0.14.0
cachetools==5.5.0
camel-ai==0.2.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cohere==4.57
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
crashtest==0.4.1
cryptography==43.0.1
cssselect==1.2.0
curl_cffi==0.6.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.19.1
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.30.3
dill==0.3.8
discord.py==2.4.0
distlib==0.3.8
distro==1.9.0
docker==7.1.0
docstring-parser==0.15
docx2txt==0.8
duckduckgo_search==6.2.12
dulwich==0.21.7
effdet==0.4.1
emoji==2.13.0
environs==9.5.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.1.0
fastavro==1.9.7
fastjsonschema==2.20.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
firecrawl-py==1.2.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.1
geojson==2.5.0
google-ai-generativelanguage==0.6.4
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
googlemaps==4.10.0
groq==0.5.0
grpcio==1.66.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
importlib-metadata==6.11.0
installer==0.7.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jedi==0.19.1
jeepney==0.8.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==24.3.1
kiwisolver==1.4.7
langdetect==1.0.9
layoutparser==0.3.4
lazy-object-proxy==1.10.0
litellm==1.46.8
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
milvus-lite==2.4.10
mistralai==1.1.0
more-itertools==10.5.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
neo4j==5.24.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.15.1
openai==1.46.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==23.2
pandas==2.2.3
pandoc==2.4
parso==0.8.4
pathable==0.4.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.3.6
plumbum==1.8.3
ply==3.11
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
portalocker==2.10.1
prance==23.6.21.0
praw==7.7.1
prawcore==2.4.0
primp==0.6.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyGithub==2.4.0
Pygments==2.18.0
PyJWT==2.9.0
pymilvus==2.4.6
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyNaCl==1.5.0
pyowm==3.3.0
pypandoc==1.13
pyparsing==3.1.4
pypdfium2==4.30.0
pyproject_hooks==1.1.0
PySocks==1.7.1
pyTelegramBotAPI==4.22.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
qdrant-client==1.11.2
rank-bm25==0.2.2
rapidfuzz==3.9.7
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
reka-api==3.0.8
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
sentence-transformers==3.1.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
slack_sdk==3.33.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
trove-classifiers==2024.9.12
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.10.30
unstructured-inference==0.7.11
unstructured.pytesseract==0.3.13
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
